requirements.txt
flask
flask-socketio
eventlet
gevent
mysql-connector-python
gunicorn
